requirements.txt
peewee~=3.17.1
configobj~=5.0.8
pydantic~=2.5.3
PyJWT~=2.8.0
pytz~=2024.1
pydantic[email]
python-multipart
paho-mqtt==1.5.1
xlrd>=1.0.0
fastapi~=0.110.0
starlette~=0.36.3
psycopg2-binary~=2.9.9
uvicorn~=0.28.0
dynaconf~=3.2.5
dependency-injector~=4.41.0
pytest~=8.1.1
httpx~=0.27.0
python-dotenv~=1.0.1
dynaconf~=3.2.5
pytest-asyncio~=0.23.6
pytest-mock~=3.14.0
python-dateutil~=2.9.0.post0